requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-chess==1.999
stockfish==3.28.0
pandas==2.1.4
requests==2.31.0

